Changed while also to be sequential, changed compiler API to better allow for recompilation, actually added destructors.

+++ Testing/auto_functions.h
@@ -24,6 +24,6 @@ bool RaiseWings(void *data);
 bool TestBool(void *data, bool willPrintMe, bool otherBool);
 
 bool PointTurn(void *data, float x, float y, bool highgear);
-bool Wait(void *data, float t);
+bool Wait(void *data, int t);
 
 #endif /* auto_functions_h */
